One more exclusive-lock
Summary:
If for some reasons, some thread is calling compute-slots,
while another is finalizing/reinitializing, then we can lead to state
of losing an index.

Can this happen in practice? Well, D6502 should've fixed the index
issues, but we immediately saw an index failure in the Gatekeeper
class, of all things.

Now the Gatekeepers were definitely not being modified at the time. It
could've been read, but it was unlikely. This leads me to believe that
multiple MOP commands were happening on different threads. e.g. my
REPL thread was reloading code, and maybe some code somewhere was
reading a gatekeeper.

I believe this should be safe. Hope we don't see a deadlock in prod. A
deadlock could potentially happen between the BKNR transaction lock
and the sharing lock's exclusive mode. For example, the datastore
code's reinitialize-instance writes a transaction
`update-instances-for-changed-class`. The good news is the deadlock is
unlikely to happen unsupervised, it's more likely to happen when we're
reinitializing classes.

Differential Revision: https://phabricator.tdrhq.com/D6503

GitOrigin-RevId: 4d2c67df2097e4407e98ec7f62d654f06165b3c2