Make accessors work without with-transaction
Summary:
In particular, this is important for our bknr.cluster
implementation. This might add a small overhead to transactions:
multiple actions in a transaction will each incur an IO cost. But the
cost is worth the simplification.

We can still create custom transactions if we want to combine multiple
actions into one.

Differential Revision: https://phabricator.tdrhq.com/D7558

GitOrigin-RevId: 98887c687d3fde6d000ac45df2c40bfd8413b8d2